trying to deal with duplicate entries p2
found the 'System.exit(0)' method that allows me to stop the program all together which is good and bad. Good because it doesn't ask for the monetary value but bad as it closes the java window the user can't do anything. so the code doesn't deal with duplicate entries well.